cache: split CacheManager out of utils into its own module
utils.py had grown into a grab-bag of unrelated concerns (time helpers,
HTML escaping, and ~380 lines of cache management). Move CacheManager into
a dedicated gracenote2epg/cache.py so each module has a single focus:
utils.py now holds only TimeUtils + HtmlUtils (102 lines). Import sites
updated; no behavioural change (XMLTV output is byte-identical).
